pyproject.toml
[tool.poetry]
name = "python-sqlserver-integration"
version = "0.1.0"
description = ""
authors = ["PROENÇA, Marco <contato.marcoproenca@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyodbc = "^5.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
